added new TagHelpers added Invoice and product aggregate to TestWebApp

+++ src/DNTFrameworkCore/Application/Models/DetailModel.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json;
 
 namespace DNTFrameworkCore.Application.Models
 {
-    public abstract class DetailModel : DetailModel<long>
+    public abstract class DetailModel : DetailModel<int>
     {
     }
